fix(runtime,workbench): route missing runtime methods, optimistic model capability discovery, workbench tool schemas
- Route agent.session.readWindow/readToolArtifact, provider.profile.saveAndDiscover,
  models.updateCapabilities through handle_agent_request (METHOD_NOT_FOUND fixes)
- Model capabilities: optimistic tool-calling default for tool-encoding protocols,
  models.dev provider key resolution (aliases, underscore normalization, suffix
  routes), tagged model id matching, and a state-load migration upgrading stale
  false flags while respecting overrides and runtime rejections
- Add input schemas for workbench tab/terminal tools (split_tabs, detach_split,
  activate/close/reorder_tab, terminal open/focus/close/move) and document
  split-group accumulation semantics
- Split file grep fallback into its own module; platform-aware terminal tests;
  CRLF-safe verification tooling; regenerate docs and third-party notices

+++ crates/lyra-computer-use-core/src/windows.rs
@@ -1974,7 +1974,7 @@ mod tests {
         let lp = pack_lparam(i32::MAX, i32::MIN);
         let x = (lp.0 as u32) & 0xFFFF;
         let y = ((lp.0 as u32) >> 16) & 0xFFFF;
-        assert_eq!(x, i16::MAX as u16);
-        assert_eq!(y, i16::MIN as u16);
+        assert_eq!(x, i16::MAX as u16 as u32);
+        assert_eq!(y, i16::MIN as u16 as u32);
     }
 }
